backend/pyproject.toml
[project]
name = "multi-agent-ai-system"
version = "1.0.0"
description = "Production-grade multi-agent AI platform with plugin architecture"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "websockets>=12.0",
    "pydantic>=2.5.0",
    "pyyaml>=6.0.1",
    "torch>=2.0.0",
    "transformers>=4.35.0",
    "accelerate>=0.24.0",
    "faster-whisper>=0.9.0",
    "pillow>=10.0.0",
    "qdrant-client>=1.7.0",
    "numpy>=1.24.0",
    "aiofiles>=23.2.1",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"
